styles Notes and modal

+++ tomanota/src/componets/Creanota.js
@@ -2,21 +2,56 @@ import React, { useState } from "react";
 import { doc, deleteDoc } from "firebase/firestore";
 import { db } from "../firebaseconfig";
 import { Modal } from "./Modal";
+import Swal from 'sweetalert2'; 
 import btndelete from "../assets/images/btndelete.png";
 import edit from "../assets/images/edit.png";
 import "./styles/Note.css";
 
 
 export const Creanota = ({ note }) => {
-  const { id, title, description } = note;
-
+  const { id, title, description, date } = note;
+  const d= date.toDate ( ).toLocaleString();
+  console.log(d);
   const [isVisible, setIsVisible] = useState(false);
   const showModal = () => setIsVisible(true);
   const hideModal = () => setIsVisible(false);
 
-  const deleteNote = async () => {
+  const deleteNote = () => {
     try {
-      await deleteDoc(doc(db, "notes", id));
+      const swalWithButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn-success',
+        cancelButton: 'btn-danger'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithButtons.fire({
+      title: "¿Deseas borrar tu nota?",
+      text: "No podras revertir esto",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Si',
+      cancelButtonText: 'No',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        swalWithButtons.fire(
+          '¡Eliminada!',
+          'La nota ha sido borrada.',
+          'success'
+        )
+        deleteDoc (doc(db, "notes", id));
+      } else if (
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithButtons.fire(
+          'Cancelar',
+          'Tu nota se ha guardado :)',
+          'error'
+        )
+      }
+    })
     } catch (error) {
       console.log(error);
     }
@@ -27,6 +62,9 @@ export const Creanota = ({ note }) => {
       <div className="note-body">
         <h4 className= "note-title">{title}</h4>
         <p className= "note-description">{description}</p>
+      </div>
+      <div>
+        <p className= "date"> modificado: {d}</p>
       </div>
       <div className="note-btns">
         <img
